+++ CHANGELOG.md
@@ -10,6 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Extract username from session via new
   `SessionKeyGenerator.get_web_auth_session_key_username` ([#290])
+* `User.get_track_scrobbles` ([#298])
+
+### Deprecated
+
+*  `User.get_artist_tracks`. Use `User.get_track_scrobbles` as a partial replacement.
+   ([#298])
 
 ## [3.0.0] - 2019-01-01
 ### Added
@@ -31,6 +37,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [Unreleased]: https://github.com/pylast/pylast/compare/v3.0.0...HEAD
 [3.0.0]: https://github.com/pylast/pylast/compare/2.4.0...3.0.0
 [2.4.0]: https://github.com/pylast/pylast/compare/2.3.0...2.4.0
+[#298]: https://github.com/pylast/pylast/issues/298
 [#290]: https://github.com/pylast/pylast/pull/290
 [#265]: https://github.com/pylast/pylast/issues/265
 [#273]: https://github.com/pylast/pylast/issues/273